chore: Add scaffolding for uploads.begin
Most of the "core logic" is not implemented because cba today : )

That said, the thing that seemed to need to be fixed was the fact that a
lot of the scaffolding in question needed to be concretised, so here it
is.

Basic idea is to keep Controllers as close to the presentation layer as
possible, and move the interesting domain logic into Services.
(Nominally more complex services with detailed startup requirements
should be provided by a Provider, if they need to hook into the app boot
process, though I'm sure ours shouldn't.)