refactor: Use an explicit GcsUrl for referencing gs:// URLs
This prevents misuse of GcloudStorageService by requiring to pass
absolute URLs to all operations and resolving them locally as needed.

This arises from the requirement to store fully-qualified `gs://` URLs
in the database associated with files and uploads such that data isn't
lost if the default bucket or location becomes changed.