feat: Implement stale/buried job cleanup
First of all, a bit on the terminology. Uploads become _stale_ 24 hours
after they are begun, in that we assume that the upload isn't likely to
become finished after that point.

That said, given GCS's peculiarities, it's still possible for an upload
to become finished after the 24 hour period elapses: if the GCS upload
was begun right before the signed URL expires, it appears that it's
still possible for the streaming upload session to persist for a week
more, so during that week the upload might become finished. Hence we
cannot drop it from the database after the 24 hours expire so that we
have a quickly perusable list of files to delete if they exist on the
GCS side.

This commit also includes the scaffolding migration for the
database-backed job queue that we'll need in other areas as well.