doc(api): [BREAKING] Adapt other changes from design doc
There were a bunch of trivia changed along the way, but this is
different enough to demand an audit and synchronized changes in both
Foxhound and Spaniel.

This gets rid of files.check_upload_progress given that it is
fundamentally incomplete; along with the indexing state, the file type
can change during indexing as well.

Issue #26 is *NOT* resolved; this will require further changes in the
actual code that cba right now.

Resolves: #25
See: #26